+++ apps/omsorgspengesoknad-refactor/src/app/i18n/index.tsx
@@ -1,0 +1,69 @@
+import { FormattedMessage, useIntl } from 'react-intl';
+import { commonMessages } from '@navikt/sif-common-core-ds/src/i18n/common.messages';
+import { typedIntlHelper } from '@navikt/sif-common-core-ds/src/utils/intlUtils';
+import { soknadMessages } from '@navikt/sif-common-soknad-ds';
+import { uiMessages } from '@navikt/sif-common-ui';
+import { componentMessages } from './componentMessages';
+import { appMessages_nb } from './nb';
+import { appMessages_nn } from './nn';
+
+export const libMessages = {
+    nb: {
+        ...commonMessages.nb,
+        ...uiMessages.nb,
+        ...soknadMessages.nb,
+    },
+    nn: {
+        ...commonMessages.nn,
+        ...uiMessages.nn,
+        ...soknadMessages.nn,
+    },
+};
+
+export const appMessages = {
+    nb: appMessages_nb,
+    nn: appMessages_nn,
+};
+
+const nb = {
+    ...libMessages.nb,
+    ...appMessages.nb,
+    ...componentMessages.nb,
+};
+
+const nn: Record<keyof typeof nb, string> = {
+    ...libMessages.nn,
+    ...appMessages.nn,
+    ...componentMessages.nn,
+};
+
+export const appAndComponentIntlMessages = {
+    nb: {
+        ...appMessages.nb,
+        ...componentMessages.nb,
+    },
+    nn: { ...appMessages.nn, ...componentMessages.nn },
+};
+
+export type AppMessageKeys = keyof typeof nb;
+
+export const useAppIntl = () => {
+    const intl = useIntl();
+    return typedIntlHelper<AppMessageKeys>(intl);
+};
+
+export type AppIntlShape = ReturnType<typeof useAppIntl>;
+
+interface AppTextProps {
+    id: AppMessageKeys;
+    values?: any;
+}
+
+export const AppText = (props: AppTextProps) => {
+    return <FormattedMessage {...props} />;
+};
+
+export const applicationIntlMessages = {
+    nb,
+    nn,
+};