+++ apps/omsorgspengesoknad-refactor/src/app/types/SøknadContextState.ts
@@ -1,0 +1,14 @@
+import { RegistrertBarn, Søker } from '@navikt/sif-common-query';
+import { SøknadRoutes } from './SøknadRoutes';
+import { Søknadsdata } from './søknadsdata/Søknadsdata';
+
+export interface SøknadContextState {
+    versjon: string;
+    søker: Søker;
+    registrerteBarn: RegistrertBarn[];
+    søknadsdata: Søknadsdata;
+    søknadRoute?: SøknadRoutes;
+    søknadSendt?: boolean;
+    børMellomlagres?: boolean;
+    isReloadingApp?: boolean;
+}