+++ apps/omsorgspengesoknad-refactor/src/app/types/søknadsdata/DeltBostedSøknadsdata.ts
@@ -1,0 +1,5 @@
+import { Vedlegg } from '@navikt/sif-common-core-ds/src/types/Vedlegg';
+
+export interface DeltBostedSøknadsdata {
+    vedlegg: Vedlegg[];
+}