+++ packages/k9-sak-innsyn-api/src/initK9SakInnsynApiClients.ts
@@ -1,7 +1,8 @@
 import { k9Sak, innsyn } from '.';
 import { initApiClient } from './utils/initApiClient';
 
-export const initK9SakInnsynApiClients = (frontendPath: string, loginURL: string) => {
+export const initK9SakInnsynApiClients = (config: { frontendPath: string; loginURL: string }) => {
+    const { frontendPath, loginURL } = config;
     initApiClient(k9Sak.client, frontendPath, loginURL);
     initApiClient(innsyn.client, frontendPath, loginURL);
 };