+++ packages/sif-common-query/src/types/validerFritekst.ts
@@ -1,7 +1,0 @@
-export type {
-    Friteksfelt,
-    ValiderFriteksfeltData,
-    ValiderFriteksfeltResponses,
-    ValiderFriteksfeltErrors,
-    ValiderFriteksfeltError,
-} from '@navikt/k9-brukerdialog-prosessering-api';